requirements.txt
# Core CLI dependencies
ddgs
beautifulsoup4
openai
httpx
requests
anthropic
google-generativeai
